VERIFICATION v0.8.0 (Punkt 3.5), GOVERNANCE v0.2.0 (Punkt 1.4): Kontrollsegment-Vorrat und Beobachtungsfenster als Governance-Parameter mit Invariante Vorrat >= Fenster * gamma (Punkt 5 der hardwarefreien Liste, Fund 58); ⚑ zweite Fassung derselben Schranke in reicht_fuer | GOVERNANCE/myl-governance/examples/schranke_meldung.rs (neu); GOVERNANCE/myl-governance/src/{registry,invarianten}.rs; GOVERNANCE/myl-governance/tests/{akzeptanz,gleichstand}.rs; VERIFICATION/myl-verifier/src/{kontrollsegmente,unterscheider,lib}.rs; VERIFICATION/myl-verifier/examples/a11_kurve.rs; {VERIFICATION,GOVERNANCE}/README/README.md; SIMULATION/Sicherheitsaudit.md; Cargo.toml (2 Crates), Cargo.lock (6)

+++ VERIFICATION/myl-verifier/examples/a11_kurve.rs
@@ -4,7 +4,7 @@ fn main() {
         let e = myl_verifier::messe_wiederholung(vorrat, 100_000, 2, 100, &[7u8; 32]);
         println!("  {:>12} | {:>20} | {:>3}.{} %",
             vorrat,
-            myl_verifier::reichweite(vorrat, 2, 100),
+            myl_verifier::reichweite(vorrat as u64, 2, 100),
             e.erkannt_promille() / 10, e.erkannt_promille() % 10);
     }
     println!("  nötiger Vorrat für 100k Aufträge bei γ=2%: {}",
